Move cert to its own module
